% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getauthorlist.R
\name{getauthorlist}
\alias{getauthorlist}
\title{get author list with all authors but not curators, in order}
\usage{
getauthorlist(panels)
}
\arguments{
\item{panels}{}
}
\value{
data.frame
}
\description{
get author list with all authors but not curators, in order
}
\examples{
getauthorlist(sdashpanels(5))
}
