% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdashpanels.R
\name{sdashpanels}
\alias{sdashpanels}
\title{Get open panels information}
\usage{
sdashpanels(group = NA)
}
\arguments{
\item{group}{}
}
\value{
data.frame
}
\description{
Get open panels information
}
\examples{
sdashpanels()
sdashpanels(5)

}
